Citizenship last values (#687)
* first advance

* second advance

* Refered accounts logic added

* UI part
